Fix retina display support
For the moment, I prefer to consider that goxel only cares about
framebuffer coordinates.  So I just make sure that the input structure
gets the proper values in main.c.

Eventually we might want to make a distinction between windows and
framebuffer coordinates, for example in order to scale the UI.