Put the mesh blocks into a hash table
That way it is fast to find a block by its position.  I did this
optimization because the add_blocks function was getting too slow when
trying to load very large images.

I could probably optimize it further by using the position of the blocks
divided by the block size for the key, since blocks are always placed at
regular distance of each other.