Show the axis arrows
This is a big change because I had to refactor the way we render lines.

To prevent some bug when moving around I put the view and projection
matrix in the renderer, and they are updated once per frame.  Not sure
if this is the best way to do it though.

A lot of projection and renderer code will need to be cleaned up.