Experimental support for marching cube render
To allow fine subdivision of the voxels, I add a constant scale to the
position (set to 8 for the moment).

The colors are still a bit ugly, I couldn't figure out how to make it
work.

Also I will need to simplify the render panel in the gui, there are too
many options now.