Test using GLUT instead of GLFW3
We can compile with glut with scons glut=1.

I do that because while glfw seems to be a bit better than glut, it
would be easier to use glut if I want to make an osx bundle, since glut
is already available on osx.  I might just remove this code at some
point if I find a way to put glfw into an osx bundle.