Use zlib to save the gox files
The good thing with zlib stream functions is that they work both for
compressed and uncompressed data, so it makes the code simpler.

One downside is that zlib streams do not support backward seek, so I
have to use a buffer for the chunks.  I set the buffer to 1Mib for the
moment, if we try to save images with big chunks, it might crash.