Add action_t struct, to support function reflexion
The idea is that we keep a hash table of function pointers and associated
meta informations, like doc string, signature, etc.  then we can
dynamically call any function with a single C api.  This is nicer for
the UI.

The tricky part is that we need to support different function
signatures, which is not easy in C.