Cleanup the action code
I am afraid the code might not work on all platforms / architecture.
Basically I declare a function pointer as:

    void (*func)(); // Empty parenthesis arguments

And then later call it with variable number of arguments (casted as long):

    func(arg1, arg2);

And I hope that C code actually matches the target function.

I need to test on different platforms to see if it works or not.  If not
I can fix it at the price of a bit more verbose code.