Fix flat rendering when we zoom out
The problem was that the bump mapping texture was not sampled correctly.

I could fix it by using opengl centroid sampling, but I don't want to
reply on this because I am not sure it works everywhere.

Instead I pass the bump texture as the coordinate that need to be added
to the face uv.  Then I can clamp the uv to 0,1, solving the
problem.

I could also save memory by passing the bump texture coordinate as a
single uint8, and then using a table in the shader.