Put all the assets as binary blob in the code
Rational: this allows to distribute the program as a single exe file,
which I find very nice.  It will also make it easier to deal with the
OSX version.

So I use a pyhton script (tool/create_assets.py) that generates
srs/assets.inl with all the binary data from the data files.  Then this
file is included into assets.c, and we have a function to get the asset
data, and an other one to list assets.