Add basic support for "antialiasing"
The shape functions need to return the signed distance betweeen the
point and the surface of the shape.  Then we can use this information to
set the voxel value properly if aliased (that is a point not totally
inside inside the shape will get a fractional alpha value).

This gives better looking marching cube rendering.