Simplify meshes management
Instead of using a backup of the original current layer mesh before we apply
the brush preview on it, we instead use a temporary mesh for the brush
preview, and we only touch the current layer mesh when we actually apply
a change to it.

I think it makes the code a bit easier to follow.