Add caching or block merge operations
Because quite often we do the same block merge (for example with the
brush tool).  We could make sure we only update the part of the image
that changes, but I think it is nicer instead to cache the merge
operation so that they get basically free.  This takes more memory, but
make the code simpler.