Try to optimize move operations
I use a cache in mesh_get_at so that it speeds up the code in the
typical case where we look for adjacent blocks.

I also changed mesh_fill to first clear the mesh, I think this makes the
number of block tested smaller.

It is still very slow!  I probably need to use the block hash instead of
an iteration.