Add USER_PASS and USER_GET macros
This can be used when we have functions taking callbacks with a void
pointer argument for user data.

When we want to pass several arguments, the 'proper' way is to bundle
them into a structure, but this makes the code quite verbose.

Those macros allow to pass the arguments in a pointer array.  It ise
used like that:

void func(void (*cb)(void *user), void *user);

void my_cb(void *user)
{
    int x = USER_GET(user, 0);
    double y = USER_GET(user, 1);
    ...
}

int main()
{
    int x = 10;
    double y = 20.0;
    func(my_cb, USER_PASS(x, y));
}