Add a global 3d cursor
Similar to the way it works in blender.  For goxel this is purely
internal (for the moment).

I do this because ultimately I want to totally dissociate the tools code
for the user mouse input.  Once this is done it will be easier to
support things like touch screen or 3d sensors.

Having a global 3d cursor also allow to show it's position in the UI.

It's a pretty big change, many small parts of the projection code had to
be changed.  There are also probably many small bugs left to fix.