Cleanup imgui_user.inl
The goal is to move all the code from imgui_user.inl to gui.cpp.

ImGui has now a better support for accessing the internal data (by
including imgui_internal.h), so there is no real need to use
imgui_user.inl anymore.

I also try to user the C version of the gui functions as much as I can,
so that in the future I can eventually split the gui code into a cpp
part and a C part.