Refactor the history list structure
I make it so that the image itself is always included in its own
history.  This makes it easier to delete an image: just delete all the
history, that would include the image itself.

I make image_copy private and renamed it to image_snap: this is not a
real copy since the image path is not copied.  So we can only use it
for history snapshot.