First draft of mesh flood fill algo
I called it mesh_select for the moment, but the name is not very good.
I created a dummy tool (extrude) to test it.  This tool will then be the
first to use the algo in order to check what voxels we are extruding.

I didn't optimize at all for the moment.