Extrude tool start to work a bit
But the code is just terrible.  I need to do a big cleanup of the mesh
and math code.  I think the representation of boxes should be redone.
The way it works now is not very convenient.

I also need to create a generic way to have tools manipulate boxes.