Start to change the block/mesh API to use vec3i
I want to use integer position everywhere instead of floats.  Floats
make the code quite tricky in many places.  Beside, the mesh should be
seen as a kind of 3d matrix.

The problem is that it will be hard to make the change without breaking
the code a lot.  I think in the long term this is the way to go though.

Eventually I might even hide the mesh internal totally.