Start to simplify mesh iterator functions
I think it is better if the iterators only give us the position, we can
then get the value with a second call if needed.

I am still not sure what would be the perfect api, because I also would
like to support multithreaded access to the mesh, and I am not sure how
this could work.