Optimize marching cube algo.
The new mesh API is slower than the previous one at the moment, I need
to figure out how to optimize it.

Meanwhile I added a function to copy a voxel region into a memory
buffer, this can be used when we need very fast read only access to the
voxels, as it is the case for the marching cube algorithm.