Remove all 'goxel' argument from functions api
I also create the global goxel instance in goxel.c instead of main.

I don't like to charge the API with an attribute that will always be a
global variable anyway.  The best would be to minimize the use of the goxel
attribute, for example the update_mesh function should be removed I
think.