mesh_generate_vertice returns some more info
Depending on the effect (marching cube or not), the array returned by
mesh_generate_vertice would represent quads or triangles, and the unit
size would be 1 or 1/8.

Until now the renderer had to manually check for that.  This lets the
vertice function declares what the values should be, as it should be.