Experimental test of running action in threads
I want to be able to show a progress indicator during long export
operations, the current code didn't allow that since the action are
always run in the main thread.

This patch add the possibility to run action in a separate thread.

In order to prevent thread issues, I add a global mutex so that the main
iter function and the actions never actually run at the same time.  The
goal is to have the action run in between the main loop iterations.
This might be too slow, and in that case we will need to actually run
the action at the same time as the main thread.

Anyway this is just a test for the moment.  I am not sure I will merge
that.