Some small changes to fix cycles on Windows
Some minor changes in cycles codes so that it compile with msys2.

In the build system I had to use -O3 minus ftree-slp-vectorize instead
of -Ofast, otherwise the rendering would segfault, with no clear reason
(bug in gcc?)

I also put -msse2, I hope this doesn't make the binary non portable!

All in all I am not very pleased with cycles, it's slow to compile and
the code doesn't play well with msys2.