Some attempts at fixing cycles compilation
I am not quite sure how it should work.  It seems that cycles always
require SSE2 support, and the compilation fails on some target ( I just
tried on a i586 virtual machine) if we don't set the -msse2 flag.

Also there is a crash when I use tree-slp-vectorize!

Long term it would be nice to replace cycles with a smaller library that
is easier to compile.