Cleanup box_edit
I realize that it is currently impossible to implement a generic box
editor that can be applied on any box.  It is only possible to edit
boxes the core can snap to, which means only the layer, selection, or
image at the moment.

Eventually we should add a way to snap on as many boxes as we want I
guess?