Disable antialiasing for add/sub modes
The problem is that at the moment the marching cube algorithm doesn't
work well with non binary voxels.  I could probably try to fix it, but I
think it's better to just disable this mode totally, since it makes
things more complicated.  I don't think anybody was actually using that.