Remove cycles support
Too many compilation and implementation problems with this lib.  I give
up.  I'll try to use a simpler ray tracer library or implement my own.

On a side note: it's really annoying that most C++ libraries seem to
need to be constantly updated or they fail to compile with new versions
of compilers.