gui: change number input code
I am trying to make the code compatible with the current release of
imgui, so that I can update.  This commit replaces my custom number
input widget by using imgui widgets as much as possible.

The difficulty is that the input text in imgui is centered in the
widget, but I want to render it with the label on the left and value on
the right.  I do the trick by setting the text color to transparent and
rendering the actual text after on top of the widget.