Remove lua library
This wasn't used anymore.  The lua scripting support wasn't implemented
properly so I think it is better to remove it totally.

My goal is now to cleanup the code to make the 'actions' less dynamic
(and error prone) and only then see if it is worth adding scripting
capabilities back.  I also think using quickjs would be a better option
than lua.