Prevent crash when we open empty image
Not sure if that is totally correct.  The issue is that the layers have
a pointer to the material that is owned by the image.  If we open an
empty image on top of an empty image, the hash does not change, and so
goxel.render_layers are not updated, and still point to the material of
the old image!

To fix that I add the material pointer in the layer hash, even though
this is not really correct I think.