Attempt to fix bug when we switch language
I am not totally sure about it.  Sometimes when I change lang the
settings popup closes and the UI doesn't react anymore.  Probably
something to do with the way ImGui traces popup from their titles.

For now I try to change the lang only after we finish the UI rendering,
but it might not be enough.