Change SNAP_ROUND behavior
Possibly breaks some tools!

The previous behavior was always rounding to inside the nearest voxel,
but usually when we snap to the volume we want to stay on top of it (and
then use a -0.5 offset to get inside if needed).

The new behavior only rounds along the tangent and binormal, but not the
normal.  This will make the selection tool a bit easier to implement I
think.