Redo undo stack code
I change the way it works: now we push into the stack at the end of an
edit, instead of before.  This will allow to not push into the undo
stack if we cancel some edition action.

I also changed the code so that we only put the snapshot in the undo
stack, and never the image itself.  This makes the code easier to
understand.