Use nativefiledialog instead of tinyfiledialog
tinyfiledialog is a bit annoying because it relies on external tools
that might not be installed on the running machine.

An other issue is that current version of zenity used on linux doesn't
support setting the default save path.

Here I replace with tinyfiledialog that seems to be more reliable.

For the linux AppImage I use the dbus backend so that it should use the
system default style.