Start to add better help hint
Still a work in progress.  Not sure if the hint for navigation are
correct.

The idea is to have different hint depending on the tool and the pressed
key so that we directly see what we can do.

I added a 3d gesture flag GESTURE3D_ALWAY_CALL so that the gesture
callback ca be called even before a gesture has begun, which allows to
set the hint there.  Could be set as default maybe?