Switch to using po/mo files for translations
As a test for now.  Would need to make sure that the speed is
acceptable.

Doing this because it makes it easier to work on the UI without having
to update the translation strings list each time I add a new label.

I am using a small custom made mo reader since I don't want to add a
dependency to gettext, and also I prefer to bundle all the translations
directly in the binary.