Fix crash if we open an empy image
On top of an empty image.

This is the second version of this fix, in the first commit I made a
mistake and hashed the matrix (mat) instead of the material (material).

The issue is that we need to make sure that the hash changes when we
open a new image, even if the image is exactly the same, because the
material pointer of the cached render_layers still point to the old
image.

I think a proper solution would be to avoid using pointer for
layer->material but instead an index.  The same could be done for
image->layer, image->camera, etc.