Add basic support for drag and drop of lists
Added a generic gui component for lists, so that we can use it for
layer, cameras and materials.  For the moment this supposes that the
list items are all based on the same structure that starts with a ref
counting and two pointers to next and prev items.

I think I should create some kind of base struct for all the lists to
avoid this though.

Since I make all the list behave the same, I changed the order of the
layers.  Hopefully this is acceptable.